+++ directio/directio_darwin.go
@@ -1,4 +1,4 @@
-package env
+package directio
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 //                                                                                    //
@@ -8,30 +8,33 @@ package env
 // ////////////////////////////////////////////////////////////////////////////////// //
 
 import (
-	"fmt"
+	"errors"
+	"os"
+	"syscall"
 )
 
 // ////////////////////////////////////////////////////////////////////////////////// //
 
-func ExampleGet() {
-	env := Get()
+const (
+	BLOCK_SIZE = 4096 // Minimal block size
+	ALIGN_SIZE = 0    // Align size
+)
 
-	// Print PATH environment variable
-	fmt.Println(env["PATH"])
+// ////////////////////////////////////////////////////////////////////////////////// //
 
-	// Path return PATH variable as slice
-	for i, p := range env.Path() {
-		fmt.Printf("%d %s\n", i, p)
+func openFile(file string, flag int, perm os.FileMode) (*os.File, error) {
+	fd, err := os.OpenFile(file, flag, perm)
+
+	if err != nil {
+		return nil, err
 	}
 
-	// You can use getters for different value formats
-	fmt.Printf("Integer value %s = %d\n", "INT_VALUE", env.GetI("INT_VALUE"))
-	fmt.Printf("Float value %s = %g\n", "FLOAT_VALUE", env.GetF("FLOAT_VALUE"))
-	fmt.Printf("String value %s = %s\n", "STR_VALUE", env.GetS("STR_VALUE"))
-}
+	_, _, e := syscall.Syscall(syscall.SYS_FCNTL, uintptr(fd.Fd()), syscall.F_NOCACHE, 1)
 
-func ExampleWhich() {
-	echoPath := Which("echo")
+	if e != 0 {
+		fd.Close()
+		return nil, errors.New("Can't set F_NOCACHE for given file")
+	}
 
-	fmt.Printf("Full path to echo binary is %s\n", echoPath)
+	return fd, nil
 }